+++ core/src/main/java/io/honeycomb/opentelemetry/android/HoneycombDeterministicSampler.kt
@@ -1,5 +1,6 @@
 package io.honeycomb.opentelemetry.android
 
+import android.util.Log
 import io.opentelemetry.api.common.Attributes
 import io.opentelemetry.api.trace.SpanKind
 import io.opentelemetry.context.Context
@@ -8,13 +9,18 @@ import io.opentelemetry.sdk.trace.samplers.Sampler
 import io.opentelemetry.sdk.trace.samplers.SamplingDecision
 import io.opentelemetry.sdk.trace.samplers.SamplingResult
 
+private val TAG = "HoneycombDeterministicS"
+
 class HoneycombDeterministicSampler(private val sampleRate: Int) : Sampler {
     private val inner: Sampler =
         if (sampleRate < 1) {
+            Log.d(TAG, "Sample rate too low, not emitting any spans!")
             Sampler.alwaysOff()
         } else if (sampleRate == 1) {
+            Log.d(TAG, "Not sampling, emitting all spans")
             Sampler.alwaysOn()
         } else {
+            Log.d(TAG, "Sampling enabled: emitting every 1 in $sampleRate spans")
             Sampler.traceIdRatioBased(1.0 / sampleRate)
         }