+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* feat: Add deterministic sampler
+* feat: Add deterministic sampler (configurable through the `sampleRate` option)
 * refactor: Move touch auto-instrumentation to separate package
 * feat: Add instrumentation helpers for Android Compose